+++ DataLayer/Event/ViewCart.php
@@ -6,18 +6,28 @@ use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Yireo\GoogleTagManager2\Api\Data\EventInterface;
 use Yireo\GoogleTagManager2\DataLayer\Tag\Cart\CartItems;
+use Yireo\GoogleTagManager2\DataLayer\Tag\Cart\CartValue;
+use Yireo\GoogleTagManager2\DataLayer\Tag\CurrencyCode;
 
 class ViewCart implements EventInterface
 {
     private CartItems $cartItems;
+    private CartValue $cartValue;
+    private CurrencyCode $currencyCode;
 
     /**
      * @param CartItems $cartItems
+     * @param CartValue $cartValue
+     * @param CurrencyCode $currencyCode
      */
     public function __construct(
-        CartItems $cartItems
+        CartItems $cartItems,
+        CartValue $cartValue,
+        CurrencyCode $currencyCode
     ) {
         $this->cartItems = $cartItems;
+        $this->cartValue = $cartValue;
+        $this->currencyCode = $currencyCode;
     }
 
     /**
@@ -31,6 +41,8 @@ class ViewCart implements EventInterface
             'cacheable' => true,
             'event' => 'view_cart',
             'ecommerce' => [
+                'currency' => $this->currencyCode->get(),
+                'value' => $this->cartValue->get(),
                 'items' => $this->cartItems->get()
             ]
         ];